statistic creation and invalidation moved to simulationResult controllers, new migration updating statistics table
