+++ backend/src/app/controllers/simulationResult.ts
@@ -6,8 +6,8 @@ import { limitQuery } from '../../lib/queryParamHandlers/limit';
 import { offsetQuery } from '../../lib/queryParamHandlers/offset';
 import { filterHandler } from '../../lib/queryParamHandlers/filterHandler';
 import { tableName } from '../../lib/tableName';
-import { SimulationResultValidity, SimulationResultStatus } from '../../lib/enums';
-import * as statistic from '../controllers/statistic';
+import { SimulationResultValidity, SimulationResultStatus, StatisticValidity } from '../../lib/enums';
+import { Statistic } from '../models/statistic';
 
 export const index = async (req: Request, res: Response) => {
   try {
@@ -67,6 +67,52 @@ const decodePayload = (req: Request): SimulationResultPayload => {
   }  
 };
 
+const convertTimeToNanosec = (measurementAsString: string): number => {  // 1234.55ms
+  const index = measurementAsString.search(/[a-zA-Z]+/);
+  let measurementValue = parseFloat(measurementAsString.substring(0, index));
+  const measurementTimeUnit = measurementAsString.substring(index, measurementAsString.length);
+  if (measurementTimeUnit === "sec") {
+    measurementValue *= 1000000000;
+  } else if (measurementTimeUnit === "min") {
+    measurementValue *= 60000000000;
+  } else if (measurementTimeUnit === "ms") {
+    measurementValue *= 1000000;
+  } else if (measurementTimeUnit === "μs") {
+    measurementValue *= 1000;
+  }
+  return measurementValue;
+};
+
+const createStatistic = async (simulationResult: SimulationResult) => {
+  try {
+    const payload_data = JSON.parse(simulationResult.payload_data);
+    console.log(payload_data);
+    const start_timestamp = simulationResult.start_timestamp;
+    for (let stage of Object.keys(payload_data)) {
+      for (let i = 0; i < payload_data[stage].length; i++) {
+        if (payload_data[stage][i].measurement !== '') {
+          const methodEndpoint = payload_data[stage][i].endpoint.split(" ");
+          const method = methodEndpoint[0];
+          const fullEndpoint = methodEndpoint[1].split("/api/");
+          const endpoint = "/api/" + fullEndpoint[1];
+          const measurement = convertTimeToNanosec(payload_data[stage][i].measurement);
+          const statistic: Statistic = {
+            simulation_result_id: simulationResult.id,
+            start_timestamp: start_timestamp,
+            method: method,
+            endpoint: endpoint,
+            measurement: measurement,
+            invalid: StatisticValidity.VALID
+          }
+          await database(tableName.STATISTICS).insert(statistic);
+        }
+      }
+    }
+  } catch (error) {
+    console.log(error);
+  }
+};
+
 const getUpdatedSimulationResult = (req: Request): SimulationResult =>{
   try {
     const decodedPayload: SimulationResultPayload = decodePayload(req);
@@ -95,9 +141,10 @@ export const update = async (req: Request, res: Response) => {
     if (initializedSimulationResult) {
       const updatedSimulationResult: SimulationResult = getUpdatedSimulationResult(req);
       await database(tableName.SIMULATION_RESULTS).update(updatedSimulationResult).where({ id: req.params.id });
-      try{
-        statistic.create(updatedSimulationResult);
+      try {
+        createStatistic(updatedSimulationResult);
       } catch (error) {
+        console.log(error);
         res.sendStatus(500);
       }
       res.sendStatus(200);
@@ -125,12 +172,23 @@ export const destroy = async (req: Request, res: Response) => {
   }
 };
 
+const invalidateStatistic = async (simulation_result: Partial<SimulationResult>) => {
+  try {
+    const statistics: Array<Statistic> = await database(tableName.STATISTICS).select().where({ simulation_result_id: simulation_result.id });
+    if (statistics) {
+      await database(tableName.STATISTICS).where({ simulation_result_id: simulation_result.id}).update({ invalid: StatisticValidity.INVALID });
+    }
+  } catch (error) {
+    console.log(error);
+  }
+};
+
 export const invalidate = async (req: Request, res: Response) => {
   try {
     const simulation_result: Partial<SimulationResult> = await database(tableName.SIMULATION_RESULTS).select().where({ id: req.params.id }).first();
     if (simulation_result) {
       await database(tableName.SIMULATION_RESULTS).where({ id: req.params.id }).update({ invalid: SimulationResultValidity.INVALID });
-      statistic.invalidate(simulation_result);
+      invalidateStatistic(simulation_result);
       res.sendStatus(200);
     } else { 
       res.sendStatus(404);
